Refactor extract download logic
Previously, cmd/download runDownload contained all the logic for getting
a downloadPayload, writing metadata, and writing solution files.

These changes separate and extract this logic outside
cmd/download for use by future commands.
