+++ cmd/download.go
@@ -1,18 +1,10 @@
 package cmd
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	netURL "net/url"
-	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 
-	"github.com/exercism/cli/api"
 	"github.com/exercism/cli/config"
 	"github.com/exercism/cli/workspace"
 	"github.com/spf13/cobra"
@@ -79,191 +71,52 @@ func runDownload(cfg config.Config, flags *pflag.FlagSet, args []string) error {
 		return err
 	}
 
-	param := "latest"
+	urlParam := "latest"
 	if uuid != "" {
-		param = uuid
+		urlParam = uuid
 	}
-	url := fmt.Sprintf("%s/solutions/%s", usrCfg.GetString("apibaseurl"), param)
 
-	client, err := api.NewClient(usrCfg.GetString("token"), usrCfg.GetString("apibaseurl"))
+	params := downloadParams{
+		cfg:      cfg,
+		uuid:     uuid,
+		slug:     slug,
+		track:    track,
+		team:     team,
+		urlParam: urlParam,
+	}
+	payload, err := getDownloadPayload(params)
 	if err != nil {
 		return err
 	}
 
-	req, err := client.NewRequest("GET", url, nil)
-	if err != nil {
+	if err := writeMetadataFromPayload(payload, cfg); err != nil {
 		return err
 	}
 
-	if uuid == "" {
-		q := req.URL.Query()
-		q.Add("exercise_id", slug)
-		if track != "" {
-			q.Add("track_id", track)
-		}
-		if team != "" {
-			q.Add("team_id", team)
-		}
-		req.URL.RawQuery = q.Encode()
-	}
-
-	res, err := client.Do(req)
-	if err != nil {
+	if err := writeSolutionFilesFromPayload(payload, cfg); err != nil {
 		return err
 	}
 
-	var payload downloadPayload
-	defer res.Body.Close()
-	if err := json.NewDecoder(res.Body).Decode(&payload); err != nil {
-		return fmt.Errorf("unable to parse API response - %s", err)
-	}
-
-	if res.StatusCode == http.StatusUnauthorized {
-		siteURL := config.InferSiteURL(usrCfg.GetString("apibaseurl"))
-		return fmt.Errorf("unauthorized request. Please run the configure command. You can find your API token at %s/my/settings", siteURL)
-	}
-
-	if res.StatusCode != http.StatusOK {
-		switch payload.Error.Type {
-		case "track_ambiguous":
-			return fmt.Errorf("%s: %s", payload.Error.Message, strings.Join(payload.Error.PossibleTrackIDs, ", "))
-		default:
-			return errors.New(payload.Error.Message)
-		}
-	}
-
-	metadata := workspace.ExerciseMetadata{
-		AutoApprove: payload.Solution.Exercise.AutoApprove,
-		Track:       payload.Solution.Exercise.Track.ID,
-		Team:        payload.Solution.Team.Slug,
-		Exercise:    payload.Solution.Exercise.ID,
-		ID:          payload.Solution.ID,
-		URL:         payload.Solution.URL,
-		Handle:      payload.Solution.User.Handle,
-		IsRequester: payload.Solution.User.IsRequester,
-	}
-
-	root := usrCfg.GetString("workspace")
-	if metadata.Team != "" {
-		root = filepath.Join(root, "teams", metadata.Team)
-	}
-	if !metadata.IsRequester {
-		root = filepath.Join(root, "users", metadata.Handle)
-	}
-
-	exercise := workspace.Exercise{
-		Root:  root,
-		Track: metadata.Track,
-		Slug:  metadata.Exercise,
-	}
-
-	dir := exercise.MetadataDir()
-
-	if err := os.MkdirAll(dir, os.FileMode(0755)); err != nil {
-		return err
-	}
-
-	err = metadata.Write(dir)
-	if err != nil {
-		return err
-	}
-
-	for _, file := range payload.Solution.Files {
-		unparsedURL := fmt.Sprintf("%s%s", payload.Solution.FileDownloadBaseURL, file)
-		parsedURL, err := netURL.ParseRequestURI(unparsedURL)
-
-		if err != nil {
-			return err
-		}
-
-		url := parsedURL.String()
-
-		req, err := client.NewRequest("GET", url, nil)
-		if err != nil {
-			return err
-		}
-
-		res, err := client.Do(req)
-		if err != nil {
-			return err
-		}
-		defer res.Body.Close()
-
-		if res.StatusCode != http.StatusOK {
-			// TODO: deal with it
-			continue
-		}
-		// Don't bother with empty files.
-		if res.Header.Get("Content-Length") == "0" {
-			continue
-		}
-
-		// TODO: if there's a collision, interactively resolve (show diff, ask if overwrite).
-		// TODO: handle --force flag to overwrite without asking.
-
-		// Work around a path bug due to an early design decision (later reversed) to
-		// allow numeric suffixes for exercise directories, allowing people to have
-		// multiple parallel versions of an exercise.
-		pattern := fmt.Sprintf(`\A.*[/\\]%s-\d*/`, metadata.Exercise)
-		rgxNumericSuffix := regexp.MustCompile(pattern)
-		if rgxNumericSuffix.MatchString(file) {
-			file = string(rgxNumericSuffix.ReplaceAll([]byte(file), []byte("")))
-		}
-
-		// Rewrite paths submitted with an older, buggy client where the Windows path is being treated as part of the filename.
-		file = strings.Replace(file, "\\", "/", -1)
-
-		relativePath := filepath.FromSlash(file)
-		dir := filepath.Join(metadata.Dir, filepath.Dir(relativePath))
-		os.MkdirAll(dir, os.FileMode(0755))
-
-		f, err := os.Create(filepath.Join(metadata.Dir, relativePath))
-		if err != nil {
-			return err
-		}
-		defer f.Close()
-		_, err = io.Copy(f, res.Body)
-		if err != nil {
-			return err
-		}
-	}
-	fmt.Fprintf(Err, "\nDownloaded to\n")
-	fmt.Fprintf(Out, "%s\n", metadata.Dir)
+	fmt.Fprintf(Err, "\nDownloaded to\n%s\n", getExerciseDirFromPayload(payload, cfg))
 	return nil
 }
 
-type downloadPayload struct {
-	Solution struct {
-		ID   string `json:"id"`
-		URL  string `json:"url"`
-		Team struct {
-			Name string `json:"name"`
-			Slug string `json:"slug"`
-		} `json:"team"`
-		User struct {
-			Handle      string `json:"handle"`
-			IsRequester bool   `json:"is_requester"`
-		} `json:"user"`
-		Exercise struct {
-			ID              string `json:"id"`
-			InstructionsURL string `json:"instructions_url"`
-			AutoApprove     bool   `json:"auto_approve"`
-			Track           struct {
-				ID       string `json:"id"`
-				Language string `json:"language"`
-			} `json:"track"`
-		} `json:"exercise"`
-		FileDownloadBaseURL string   `json:"file_download_base_url"`
-		Files               []string `json:"files"`
-		Iteration           struct {
-			SubmittedAt *string `json:"submitted_at"`
-		}
-	} `json:"solution"`
-	Error struct {
-		Type             string   `json:"type"`
-		Message          string   `json:"message"`
-		PossibleTrackIDs []string `json:"possible_track_ids"`
-	} `json:"error,omitempty"`
+func getExerciseDirFromPayload(payload *downloadPayload, cfg config.Config) string {
+	usrCfg := cfg.UserViperConfig
+
+	root := usrCfg.GetString("workspace")
+	if payload.Solution.Team.Slug != "" {
+		root = filepath.Join(root, "teams", payload.Solution.Team.Slug)
+	}
+	if !payload.Solution.User.IsRequester {
+		root = filepath.Join(root, "users", payload.Solution.User.Handle)
+	}
+	exercise := workspace.Exercise{
+		Root:  root,
+		Track: payload.Solution.Exercise.Track.ID,
+		Slug:  payload.Solution.Exercise.ID,
+	}
+	return exercise.MetadataDir()
 }
 
 func setupDownloadFlags(flags *pflag.FlagSet) {
